Add khanify response
Hubot should respond to 'khanify <term>' with a khaaaan picture, with
the term khaaaaanified.

Essentially I duplicate the first vowel (except I) in the string by 6
and add an exclamation mark and then pass it to memecaptain.

If there's no vowel, add more exclamation marks so it still seems
khanified.